chore: wip refactor avatar
